# openapi.json
{"openapi":"3.1.0","info":{"title":"Kasparro Backend Assessment","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/data":{"get":{"summary":"Get Data","operationId":"get_data_data_get","parameters":[{"required":false,"schema":{"type":"string","title":"Source"},"name":"source","in":"query"},{"required":false,"schema":{"type":"integer","title":"Page","default":1},"name":"page","in":"query"},{"required":false,"schema":{"type":"integer","title":"Page Size","default":50},"name":"page_size","in":"query"},{"required":false,"schema":{"type":"string","title":"Q"},"name":"q","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ItemOut"},"type":"array","title":"Response Get Data Data Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/stats":{"get":{"summary":"Stats","operationId":"stats_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/run-etl":{"post":{"summary":"Run Etl Endpoint","operationId":"run_etl_endpoint_run_etl_post","parameters":[{"required":false,"schema":{"type":"string","title":"X Api Key"},"name":"x_api_key","in":"query"},{"required":false,"schema":{"type":"string","title":"X-Api-Key"},"name":"x-api-key","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"db":{"type":"boolean","title":"Db"},"etl_last_run":{"type":"object","title":"Etl Last Run"}},"type":"object","required":["db"],"title":"HealthResponse"},"ItemOut":{"properties":{"id":{"type":"integer","title":"Id"},"source":{"type":"string","title":"Source"},"external_id":{"type":"string","title":"External Id"},"title":{"type":"string","title":"Title"},"content":{"type":"string","title":"Content"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","source","external_id"],"title":"ItemOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}